refactor: rename observers folder to features
