+++ src/chrome-scripts/features/homeworkPercent.js
