+++ src/chrome-scripts/features/readingTime.js
