+++ src/chrome-scripts/features/searchButton.js
