+++ src/chrome-scripts/features/webinarPercent.js
